[#1596] move functions/classes & remove test files
