--- src/openzaak/components/documenten/tests/import/files/import.csv
@@ -1,4 +1,0 @@
-"uuid","identificatie","bronorganisatie","creatiedatum","titel","vertrouwelijkheidaanduiding","auteur","status","formaat","taal","bestandsnaam","bestandsomvang","bestandspad","link","beschrijving","indicatieGebruiksrecht","verschijningsvorm","ondertekening.soort","ondertekening.datum","integriteit.algoritme","integriteit.waarde","integriteit.datum","informatieobjecttype","zaakId","trefwoorden"
-,"identificatie-123",111222333,2019-08-24,"document ZXY","intern","organisatie-123","definitief","application/pdf","dut","foobar.doc",8192,"/tmp/foobar.docx",,"zeer mooi bestand","true","verschijningsvorm","analoog",2019-08-24,"crc_16","waarde",2019-08-24,"http://testserver/catalogi/api/v1/informatieobjecttypen/2eefe81d-8638-4306-b079-74e4e41f557b","d8ebc179-6de7-4400-a0bb-6f9e336d692e","samenvatting,2019,pdf"
-,,111222333,2019-08-24,"document XYZ",,"organisatie-123",,,"dut",,,"/tmp/foobar.docx",,,,,,,,,,"http://testserver/catalogi/api/v1/informatieobjecttypen/2eefe81d-8638-4306-b079-74e4e41f557b","d8ebc179-6de7-4400-a0bb-6f9e336d692e",
-,,111222333,2019-08-24,"document YZX",,"organisatie-123",,,"eng",,,"/tmp/foobar.pdf",,,,,,,,,,"http://testserver/catalogi/api/v1/informatieobjecttypen/2eefe81d-8638-4306-b079-74e4e41f557b","d8ebc179-6de7-4400-a0bb-6f9e336d692e",
