--- src/openzaak/components/documenten/tests/import/files/import_missing_headers.csv
@@ -1,1 +1,0 @@
-identificatie,bronorganisatie,creatiedatum,titel,vertrouwelijkheidaanduiding,auteur,status,formaat,taal,bestandsnaam,bestandsomvang,bestandspad,link,beschrijving,indicatieGebruiksrecht,verschijningsvorm,ondertekening.soort,ondertekening.datum,integriteit.algoritme,integriteit.waarde,integriteit.datum,informatieobjecttype
